ref(vac): refactor arquitectura
