+++ projects/portal/src/app/components/portal-menu/portal-menu.component.ts
@@ -1,5 +1,6 @@
 import { Component, ElementRef } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
+
 @Component({
     selector: 'pdp-menu',
     templateUrl: './portal-menu.component.html'
@@ -9,7 +10,8 @@ export class PDPMenuComponent {
 
     constructor(
         private el: ElementRef,
-        private router: Router
+        private router: Router,
+        private activeRoute: ActivatedRoute
     ) { }
 
     isResponsive() {
@@ -17,26 +19,36 @@ export class PDPMenuComponent {
         return this.width >= 980;
     }
 
-    cards = [{
-        id: 12,
-        nombre: 'turnos',
-        tipo: 'info',
-        semanticTag: 'solicitud',
-        icono: 'turno-bold',
-        path: 'mis-turnos',
-        color: '#0070cc',
-        outlet: 'listado',
-    },
-    {
-        id: 13,
-        nombre: 'certificados',
-        tipo: 'info',
-        semanticTag: 'elementos de registro',
-        icono: 'documentos',
-        path: 'mis-certificados',
-        color: '#0070cc',
-        outlet: 'listado',
-    }];
+    cards = [
+        {
+            id: 10,
+            nombre: 'relaciones',
+            tipo: 'info',
+            semanticTag: 'solicitud',
+            icono: 'familia',
+            path: 'mis-familiares',
+            color: '#0070cc',
+            outlet: 'listado'
+        }, {
+            id: 12,
+            nombre: 'turnos',
+            tipo: 'info',
+            semanticTag: 'solicitud',
+            icono: 'turno-bold',
+            path: 'mis-turnos',
+            color: '#0070cc',
+            outlet: 'listado',
+        },
+        {
+            id: 13,
+            nombre: 'certificados',
+            tipo: 'info',
+            semanticTag: 'elementos de registro',
+            icono: 'documentos',
+            path: 'mis-certificados',
+            color: '#0070cc',
+            outlet: 'listado',
+        }];
 
     goTo(path) {
         this.router.navigate([path]);
