--- projects/portal/src/app/pages/portal-paciente/portal-paciente-main/mis-familiares.html
@@ -1,18 +1,0 @@
-<plex-title size="sm" titulo="Mis relaciones">
-    <plex-button type="success" icon="account-plus" tooltip="agregar una relacion" class="ml-2" size="sm">
-    </plex-button>
-    <plex-button type="warning" label="delegar cuenta" tooltip="delegar cuenta" class="ml-2 mr-3" size="sm">
-    </plex-button>
-</plex-title>
-<plex-grid size="sm" type="full">
-    <plex-card type="default" size="md" *ngFor="let familiar of familiares$ | async" (click)="selected(familiar)"
-               [selected]="familiar?.id === selectedId">
-        <img [mpiFotoPaciente]="familiar">
-        <plex-label size="lg" direction="column" titulo="{{ familiar.nombre }} {{ familiar.apellido }}"
-                    subtitulo="{{ familiar.documento }}">
-        </plex-label>
-        <plex-badge type="success">{{ familiar.relacion.nombre }}</plex-badge>
-        <plex-button type="warning" size="sm" icon="pencil"></plex-button>
-        <plex-button type="danger" size="sm" icon="account-off"></plex-button>
-    </plex-card>
-</plex-grid>
